pyproject.toml
[tool.poetry]
name = "python-roborock"
version = "0.9.0"
description = "A package to control Roborock vacuums."
authors = ["humbertogontijo <humbertogontijo@users.noreply.github.com>"]
license = "GPL-3.0-only"
readme = "README.md"
repository = "https://github.com/humbertogontijo/python-roborock"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
]
packages = [{include = "roborock"}]

[tool.poetry.scripts]
roborock = "roborock.cli:main"

[tool.poetry.dependencies]
python = "^3.9"
click = ">=8"
aiohttp = "^3.8.2"
async-timeout = "*"
pycryptodome = "~3.17.0"
pycryptodomex = {version = "~3.17.0", markers = "sys_platform == 'darwin'"}
paho-mqtt = "~1.6.1"
dacite = "~1.8.0"
construct = "^2.10.68"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dev-dependencies]
pytest-asyncio = "*"
pytest = "*"
pre-commit = "^3.2.2"
mypy = "*"
ruff = "*"
isort = "*"
black = "*"
codespell = "*"
pyshark = "^0.6"

[tool.semantic_release]
branch = "main"
version_toml = "pyproject.toml:tool.poetry.version"
build_command = "pip install poetry && poetry build"

[tool.ruff]
ignore = ["F403", "E741"]
line-length = 120

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
